# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HustleCoin Telegram Mini App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: Get all available tasks
      parameters:
        - name: telegramId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of tasks with completion status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskWithStatus"
  /tasks/{taskId}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Mark a task as completed
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCompleteRequest"
      responses:
        "200":
          description: Task completion submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCompletion"
        "400":
          description: Already completed or invalid
...
  /admin/tasks:
    post:
      operationId: createTask
      tags: [admin]
      summary: Create a new task (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "403":
          description: Forbidden
  /admin/tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [admin]
      summary: Update a task (admin)
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "403":
          description: Forbidden
  /admin/tasks/{taskId}/approve:
    post:
      operationId: approveTaskCompletion
      tags: [admin]
      summary: Approve a task completion and reward HP
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveTaskRequest"
      responses:
        "200":
          description: Task approved and HP rewarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCompletion"
        "403":
          description: Forbidden
components:
  schemas:
...
    Task:
      type: object
      required: [id, title, description, reward, status]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        reward:
          type: integer
        link:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
    TaskWithStatus:
      type: object
      required: [id, title, description, reward, status, completed, approved]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        reward:
          type: integer
        link:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        completed:
          type: boolean
        approved:
          type: boolean
        completedAt:
          type: ["string", "null"]
    TaskInput:
      type: object
      required: [title, description, reward, adminTelegramId]
      properties:
        title:
          type: string
        description:
          type: string
        reward:
          type: integer
        link:
          type: ["string", "null"]
        adminTelegramId:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        reward:
          type: integer
        link:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive]
        adminTelegramId:
          type: string
    TaskCompleteRequest:
      type: object
      required: [telegramId]
      properties:
        telegramId:
          type: string
    TaskCompletion:
      type: object
      required: [id, taskId, telegramId, approved, completedAt]
      properties:
        id:
          type: integer
        taskId:
          type: integer
        telegramId:
          type: string
        approved:
          type: boolean
        completedAt:
          type: string
    ApproveTaskRequest:
      type: object
      required: [telegramId, adminTelegramId]
      properties:
        telegramId:
          type: string
        adminTelegramId:
          type: string
...