# lib/api-zod/src/generated/types/taskCompletion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HustleCoin Telegram Mini App API
 * OpenAPI spec version: 0.1.0
 */

export interface TaskCompletion {
  id: number;
  taskId: number;
  telegramId: string;
  approved: boolean;
  completedAt: string;
}